zeroth_law_pkg/requirements-dev.txt
annotated-types==0.7.0
anyio==4.9.0
astroid==3.3.9
attrs==25.3.0
authlib==1.5.2
autopep8==2.3.2
bandit==1.8.3
black==25.1.0
build==1.2.2.post1
cachecontrol==0.14.2
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.1
cleo==2.1.0
click==8.1.8
click-default-group==1.2.4
cloudpickle==3.1.1
colorama==0.4.6
coverage==7.8.0
crashtest==0.4.1
cryptography==44.0.2
deptry==0.23.0
dill==0.4.0
distlib==0.3.9
dparse==0.6.4
dulwich==0.22.8
fastjsonschema==2.21.1
filelock==3.12.4
findpython==0.6.3
flake8==7.2.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
identify==2.6.9
idna==3.10
iniconfig==2.1.0
installer==0.7.0
isort==6.0.1
jaraco-classes==3.4.0
jaraco-context==6.0.1
jaraco-functools==4.1.0
jeepney==0.9.0
jinja2==3.1.6
jq==1.8.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
junit-xml==1.8
keyring==25.6.0
linkify-it-py==2.0.3
mando==0.7.1
markdown-it-py==3.0.0
markupsafe==3.0.2
marshmallow==4.0.0
mccabe==0.7.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
more-itertools==10.6.0
msgpack==1.1.0
mutmut==3.2.3
mypy==1.15.0
mypy-extensions==1.0.0
nodeenv==1.9.1
numpy==2.2.4
nvidia-ml-py==12.570.86
packaging==24.2
parso==0.8.4
pathspec==0.12.1
pbr==6.1.1
pbs-installer==2025.4.9
pip==25.0.1
pipdeptree==2.26.0
pkginfo==1.12.1.2
platformdirs==4.3.7
pluggy==1.5.0
poetry==2.1.2
poetry-core==2.1.2
pre-commit==4.2.0
psutil==6.0.0
py-spy==0.4.0
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.11.3
pydantic-core==2.33.1
pydocstyle==6.3.0
pyflakes==3.3.2
pygments==2.19.1
pylint==3.3.6
pyproject-hooks==1.2.0
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
pyyaml==6.0.2
radon==6.0.1
rapidfuzz==3.13.0
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
requirements-parser==0.11.0
rich==14.0.0
rpds-py==0.24.0
ruamel-yaml==0.18.10
ruff==0.11.6
safety==3.2.9
safety-schemas==0.0.5
scalene==1.5.51
secretstorage==3.3.3
setproctitle==1.3.5
setuptools==78.1.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
snowballstemmer==2.2.0
sqlite-fts4==1.0.3
sqlite-utils==3.38
stevedore==5.4.1
structlog==25.2.0
tabulate==0.9.0
textual==3.1.0
toml==0.10.2
tomlkit==0.13.2
trove-classifiers==2025.4.11.15
typer==0.15.2
types-setuptools==78.1.0.20250329
typing-extensions==4.13.2
typing-inspection==0.4.0
uc-micro-py==1.0.3
urllib3==2.4.0
uv==0.6.16
virtualenv==20.30.0
vulture==2.14
wheel==0.45.1
xdg==6.0.0
yamllint==1.37.0
zstandard==0.23.0
